Rename 'filtre_frases' script to 'filter_phrases'
